Rename file: API.swift -> Session.swift
